apps/ai_server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
openai==2.32.0
httpx==0.27.2
jiter==0.14.0
python-dotenv==1.0.1
supabase>=2.10.0
